refactor: split discovery responsibilities
